% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abscol.R
\name{abs.col}
\alias{abs.col}
\title{format columns with absolute value and custom digits}
\usage{
\method{abs}{col}(x, digits)
}
\arguments{
\item{x}{data.frame}

\item{digits}{list}
}
\value{
out
}
\description{
format columns with absolute value and custom digits
}
\examples{
set.seed(123)
df <- data.frame(u=rnorm(10), v=runif(10, 5,10))
#out <- abs.col(df, c(4,2))

}
