% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abscol.R
\name{absCol}
\alias{absCol}
\title{format columns with absolute value and custom digits}
\usage{
absCol(x, digits)
}
\arguments{
\item{x}{data.frame}

\item{digits}{list}
}
\value{
out
}
\description{
format columns with absolute value and custom digits
}
\examples{
set.seed(123)
df <- data.frame(u=rnorm(10), v=runif(10, 5,10))
out <- absCol(df, c(4,2))


}
