% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xls_as_xlsx.R
\name{convert_xls_to_xlsx}
\alias{convert_xls_to_xlsx}
\title{Helper function: convert_xls_to_xlsx}
\usage{
convert_xls_to_xlsx(exe, xls, xlsx, i, n_files, dbg = TRUE)
}
\arguments{
\item{exe}{exe}

\item{xls}{xls}

\item{xlsx}{xlsx}

\item{i}{i}

\item{n_files}{n_files}

\item{dbg}{debug (default: TRUE)}
}
\description{
Helper function: convert_xls_to_xlsx
}
