% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xls_as_xlsx.R
\name{delete_registry}
\alias{delete_registry}
\title{Helper function: delete_registry}
\usage{
delete_registry(office_folder = safe_office_folder(), dbg = TRUE)
}
\arguments{
\item{office_folder}{office folder path (default: \code{safe_office_folder})}

\item{dbg}{debug (default: TRUE)}
}
\description{
Helper function: delete_registry
}
