% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblock.R
\name{get.block}
\alias{get.block}
\title{block the data frame with interval}
\usage{
get.block(dt, n.row)
}
\arguments{
\item{dt}{data.frame}

\item{n.row}{integer}
}
\value{
out
}
\description{
block the data frame with interval
}
\examples{
set.seed(123)
df <- data.frame(u=rnorm(10), v=runif(10, 5,10))
out <- get.block(df, 4)

}
