% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsign.R
\name{get.sign}
\alias{get.sign}
\title{get the plus and minus symbols}
\usage{
get.sign(df)
}
\arguments{
\item{df}{numeric}
}
\value{
out
}
\description{
get the plus and minus symbols
}
\examples{
set.seed(123)
v <- rnorm(10)
get.sign(v)


}
