% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_credit.R
\name{get_bounds}
\alias{get_bounds}
\title{Get bounds of specified limits criteria for different ranks.}
\usage{
get_bounds(ls.st, band.lwr, band.upr)
}
\arguments{
\item{ls.st}{character vector. students ranks names c("pass", "mid","good", "elite")}

\item{band.lwr}{integer vector. the length must be same as \code{band.upr}}

\item{band.upr}{integer vector. the length must be same as \code{band.upr}}
}
\value{
tibble
}
\description{
Get bounds of specified limits criteria for different ranks.
}
\examples{
st_rank <- c("pass", "mid","good", "elite","mid","pass")
lwr_base <- c(75,80,90,96)
upr_base <- c(79,89,95,98)
tbl_bounds <- get_bounds(ls.st = st_rank,
                         band.lwr = lwr_base,
                         band.upr = upr_base)
}
