% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xls_as_xlsx.R
\name{get_excelcnv_exe}
\alias{get_excelcnv_exe}
\title{Helper function: get_excelcnv_exe}
\usage{
get_excelcnv_exe(office_folder = safe_office_folder())
}
\arguments{
\item{office_folder}{office folder path (default: \code{safe_office_folder}}
}
\value{
path containing 'excelcnv.exe'
}
\description{
Helper function: get_excelcnv_exe
}
