% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_credit.R
\name{get_score}
\alias{get_score}
\title{Generate score items to satisfy your conditions.}
\usage{
get_score(ls_st, lwr_base, upr_base, lwr_lease, upr_lease, seeds, wt)
}
\arguments{
\item{ls_st}{character vector. students ranks names c("pass", "mid","good", "elite")}

\item{lwr_base}{integer vector. the length must be same as \code{upr_base}}

\item{upr_base}{integer vector. the length must be same as \code{lwr_base}}

\item{lwr_lease}{integer vector. the length must be same as \code{upr_lease}}

\item{upr_lease}{integer vector. the length must be same as \code{lwr_lease}}

\item{seeds}{integer vector.  random sample seeds.}

\item{wt}{number vector. weights which the sum equal to 1}
}
\value{
tibble
}
\description{
Generate score items to satisfy your conditions.
}
\examples{
myst <- c("pass", "mid","good", "elite","mid","pass")
lwr0 <- c(75,80,90,96)
upr0 <- c(79,89,95,98)
lwr1 <- c(70,77,87,93)
upr1 <- c(82,92,97,99)
myseeds <- 2341
mywt <- c(0.2, 0.8)

tbl_score <- get_score(
  ls_st = myst,
  lwr_base = lwr0,
  upr_base = upr0,
  lwr_lease = lwr1,
  upr_lease = upr1,
  seeds = myseeds,
  wt = mywt)
}
