% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_nbsp.R
\name{give_nbsp}
\alias{give_nbsp}
\title{Give needed 'nbsp' symbols, and auto split with nice alignment.}
\usage{
give_nbsp(n_trim, len = 50, ratio = 4)
}
\arguments{
\item{n_trim}{number, number of Chinese characters need to  trim.}

\item{len}{number, specified total length of one line. The default is "len=50".}

\item{ratio}{integer, ratio of one Chinese characters to one nbsp. The default is "len=4".}
}
\value{
list
}
\description{
Give needed 'nbsp' symbols, and auto split with nice alignment.
}
\examples{

give_nbsp(n_trim = c(10,5))
}
