% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pf.R
\name{lm.pf}
\alias{lm.pf}
\title{Function to extract the overall ANOVA p-value out of a linear model object}
\usage{
lm.pf(modelobject)
}
\arguments{
\item{modelobject}{model object
default lm object}
}
\value{
out
}
\description{
Function to extract the overall ANOVA p-value out of a linear model object
}
