% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxest.R
\name{lx.est}
\alias{lx.est}
\title{Write latex math equation of lm estimation for rmarkdown file}
\usage{
lx.est(
  lm.mod,
  lm.dt,
  style = "srf",
  obs = "i",
  opt = c("s", "t"),
  lm.label = NULL
)
}
\arguments{
\item{lm.mod}{character}

\item{lm.dt}{data.frame}

\item{style}{character}

\item{obs}{character}

\item{opt}{character}

\item{lm.label}{character}
}
\value{
out
}
\description{
Write latex math equation of lm estimation for rmarkdown file
}
\examples{

lx_out<- lx.est(lm.mod = mod_origin, lm.dt = mroz)

mroz <- wooldridge::mroz \%>\%
as_tibble() \%>\%
  select(lwage, educ,exper, fatheduc,motheduc,everything()) \%>\%
             filter(!is.na(wage))

 mod_origin <- formula(lwage ~ educ + nwifeinc +exper+I(exper^2) + I(exper^2*city)  )
 ols_origin <- lm(formula = mod_origin, data = mroz)

lx_out<- lx.est(lm.mod = mod_origin, lm.dt = mroz)


}
