% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxpsm.R
\name{lx.psm}
\alias{lx.psm}
\title{Write latex math equation for rmarkdown file}
\usage{
lx.psm(
  x,
  y = "Y",
  intercept = TRUE,
  greek.g = c("beta"),
  greek.n,
  type = "prm",
  lm.label = NULL,
  obs = "i",
  n.row
)
}
\arguments{
\item{x}{character}

\item{y}{character}

\item{intercept}{logical}

\item{greek.g}{character}

\item{greek.n}{integer}

\item{type}{character}

\item{lm.label}{character}

\item{obs}{character}

\item{n.row}{integer}
}
\value{
out
}
\description{
Write latex math equation for rmarkdown file
}
\examples{
X <- c(paste0(rep(c("X","Z"),each=4),1:4), "fathedu", "mothedu")
Y <- "lwage"
Greek.g <- c("alpha","beta","lambda")
Greek.n <- c(4,4,2)
Greek.n <- c(4,4,2)
Obs <- "i"
N.row <- 5
Cst <- FALSE

out <- lx.psm(x =X, y = Y,greek.g = Greek.g, greek.n = Greek.n,
  type = "prm", intercept = Cst , lm.label = "prm",
  obs = Obs, n.row = N.row )

}
