% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_round}
\alias{num_round}
\title{formatC for convenient}
\usage{
num_round(num, digits = 4)
}
\arguments{
\item{num}{numbers}

\item{digits}{integer
default digits = 4}
}
\value{
out
}
\description{
formatC for convenient
}
\examples{

x <- 3.14159
out <- num_round(x)

}
