% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xls_as_xlsx.R
\name{safe_office_folder}
\alias{safe_office_folder}
\title{Helper function: safe_office_folder}
\usage{
safe_office_folder(office_path = "C:/Program Files/Microsoft Office")
}
\arguments{
\item{office_path}{office folder path (default:
"C:/Program Files (x86)/Microsoft Office")}
}
\value{
path of office folder (if existing)
}
\description{
Helper function: safe_office_folder
}
