% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmerit-package.R
\docType{package}
\name{xmerit-package}
\alias{xmerit}
\alias{xmerit-package}
\title{xmerit: Easy For Statistic And Econometrics Teaching}
\description{
Easy to show latex symbols and equations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://huhuaping.github.io/xmerit/}
}

}
\author{
\strong{Maintainer}: huaping hu \email{huhuaping01@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
