% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxest.R
\name{lx.est}
\alias{lx.est}
\title{Write latex math equation of lm estimation for rmarkdown file}
\usage{
lx.est(
  lm.mod,
  lm.dt,
  style = "srf",
  lm.n = 3,
  obs = "i",
  opt = c("s", "t"),
  inf = c(""),
  digits = c(2, 4, 2, 4),
  lm.label = NULL,
  lm.tag = NULL,
  no_dollar = FALSE
)
}
\arguments{
\item{lm.mod}{character.
you should use \code{formula()} function}

\item{lm.dt}{data.frame}

\item{style}{character.
equation style on c("srf", "srm"),
with default value style="srf"}

\item{lm.n}{integer.
numbers of independent vars of each row in the right equation.
default value lm.n = 3}

\item{obs}{character.
lower script for variables on c("i", "t"),
with default value obs ="i"}

\item{opt}{character.
list of "soft" option on c("s", "t", "p"),
with the default value opt=c("s", "t")}

\item{inf}{character.
list of "soft" option on c("over","fit","Ftest"),
with the default value opt=c("")}

\item{digits}{integer.
list of digits specification on coef result,
with the default value digits=c(2,4,2,4),
respectively to c("c","s", "t", "p")}

\item{lm.label}{character. Options for equation label,
default value NULL}

\item{lm.tag}{character. Options for equation tag,
default value "NULL".}

\item{no_dollar}{Logistic value. The equation environment
should contains double dollars,  with default value "no_dollar = FALSE"}
}
\value{
out
}
\description{
Write latex math equation of lm estimation for rmarkdown file
}
\examples{
require("magrittr")
require("wooldridge")

mroz_new <- wooldridge::mroz \%>\%
  tibble::as_tibble() \%>\%
  dplyr::select(tidyselect::all_of(c("lwage", "educ", "exper", "fatheduc","motheduc")),
    tidyselect::everything()) \%>\%
    dplyr::filter(!is.na(lwage))

mod_origin <- formula(lwage ~ educ + nwifeinc + exper + I(exper^2) + I(exper^2*city))

lx_out <- lx.est(lm.mod = mod_origin, lm.dt = mroz_new)

lx_out2 <- lx.est(lm.mod = mod_origin, lm.dt = mroz_new,
  style = c('srm'),inf = c('over','fit','Ftest'),
  lm.label = 'test-srm')

}
