% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxest.R
\name{lx.est}
\alias{lx.est}
\title{Write latex math equation of lm estimation for rmarkdown file}
\usage{
lx.est(
  lm.mod,
  lm.dt,
  style = "srf",
  obs = "i",
  opt = c("s", "t"),
  digits = c(2, 4, 2, 4),
  lm.label = NULL
)
}
\arguments{
\item{lm.mod}{character
you should use \code{formula()} function}

\item{lm.dt}{data.frame}

\item{style}{character
equation style on c("srf", "srm"),
with default value style="srf"}

\item{obs}{character
lower script for variables on c("i", "t"),
with default value obs ="i"}

\item{opt}{character
list of "soft" option on c("s", "t", "p"),
with the default value opt=c("s", "t")}

\item{digits}{integer
list of digits specification on coef result,
with the default value digits=c(2,4,2,4),
respectively to c(""c,"s", "t", "p")}

\item{lm.label}{character
default value NULL}
}
\value{
out
}
\description{
Write latex math equation of lm estimation for rmarkdown file
}
\examples{
require("magrittr")
require("wooldridge")

mroz_new <- wooldridge::mroz \%>\%
  tibble::as_tibble() \%>\%
  dplyr::select(tidyselect::all_of(c("lwage", "educ", "exper", "fatheduc","motheduc")),
    tidyselect::everything()) #\%>\%
  #filter(!is.na(dplyr::vars("wage")))

mod_origin <- formula(lwage ~ educ + nwifeinc + exper + I(exper^2) + I(exper^2*city))

lx_out <- lx.est(lm.mod = mod_origin, lm.dt = mroz_new)


}
