% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxest.R
\name{lx.est}
\alias{lx.est}
\title{Write latex math equation of lm estimation for rmarkdown file}
\usage{
lx.est(
  lm.mod,
  lm.dt,
  style = "srf",
  obs = "i",
  opt = c("s", "t"),
  lm.label = NULL
)
}
\arguments{
\item{lm.mod}{character}

\item{lm.dt}{data.frame}

\item{style}{character}

\item{obs}{character}

\item{opt}{character}

\item{lm.label}{character}
}
\value{
out
}
\description{
Write latex math equation of lm estimation for rmarkdown file
}
\examples{
require("magrittr")
require("wooldridge")

mroz_new <- wooldridge::mroz \%>\%
  tibble::as_tibble() \%>\%
  dplyr::select(tidyselect::all_of(c("lwage", "educ", "exper", "fatheduc","motheduc")),
    tidyselect::everything()) #\%>\%
  #filter(!is.na(dplyr::vars("wage")))

mod_origin <- formula(lwage ~ educ + nwifeinc + exper + I(exper^2) + I(exper^2*city))

lx_out <- lx.est(lm.mod = mod_origin, lm.dt = mroz_new)


}
